Rename '_shared' to 'Common' and adjust paths accordingly
